Refactor to support SSR.
To support visitors that do not have JS enabled, we need to either pre-render a
bunch of static content or support server-side rendering (SSR). I opted for the
latter.

As part of the update, several directories have been shifted around.

- src/ -> client/js/
- static/css/ -> client/css/
- static/ -> public/ (for non-css static content)

Additionally build.ts and dev.ts have been replaced with the frontend/
code.